requirements.txt
Django==2.2.2
djangorestframework==3.9.4
Pillow==6.0.0
psycopg2-binary==2.8.3
pytz==2019.1
sqlparse==0.3.0
